+++ plinth/modules/minetest/forms.py
@@ -16,7 +16,7 @@
 #
 
 """
-Forms for miinetest module.
+Forms for minetest module.
 """
 
 from django import forms
@@ -24,27 +24,30 @@ from django.utils.translation import ugettext_lazy as _
 
 from plinth.forms import ServiceForm
 
+
 class MinetestForm(ServiceForm):
     """Minetest configuration form"""
     max_players = forms.IntegerField(
         label=_('Maximum number of players'),
         required=True,
         help_text=_('You can change the maximum number of players playing \
-                    minetest at a single instance of time'))
+                     minetest at a single instance of time'))
 
     creative_mode = forms.BooleanField(
         label=_('Enable creative mode'),
         required=False,
-        help_text=_('Creative Mode changes the rules of the game to make it\
-                    more suitable for creative gameplay, rather than challenging\
-                    “survival” gameplay.'))
+        help_text=_('Creative mode changes the rules of the game to make it \
+                     more suitable for creative gameplay, rather than \
+                     challenging "survival" gameplay.'))
 
     enable_pvp = forms.BooleanField(
-        label=_('Enable pvp'),
+        label=_('Enable PVP'),
         required=False,
-        help_text=_('Enabling Player Vs Player will cause players to damage \
+        help_text=_('Enabling Player Vs Player will allow players to damage \
                     other players'))
 
     enable_damage = forms.BooleanField(
         label=_('Enable damage'),
-        required=False)
+        required=False,
+        help_text=_('When disabled, players cannot die or receive damage of \
+                    any kind'))